requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
